backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sentence-transformers==3.0.1
pypdf==4.3.1
python-multipart==0.0.9
numpy==1.26.4
huggingface-hub==0.25.2
# Firebase dependencies
firebase-admin==6.5.0
google-cloud-firestore==2.14.0
# Gemini Developer API (free, no billing required on Spark plan)
google-generativeai==0.7.2
requests==2.31.0
# Pinecone vector database (free tier available)
pinecone-client==5.0.1

